% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryseq_function.R
\name{dryseq}
\alias{dryseq}
\title{A Cat Function}
\usage{
dryseq(
  countData,
  group,
  time,
  T_ = 24,
  sample_name = colnames(countData),
  batch = rep("A", length(sample_name)),
  n.cores = round(detectCores() * 0.6, 0)
)
}
\arguments{
\item{countData}{matrix containing non-negative integers; each column represents a sample, each row represents a gene/transcript.}

\item{group}{vector containing the name of each sample.}

\item{time}{vector containing numeric values of the time for each sample.}

\item{T_}{numeric value to indicate period length of the oscillation.}

\item{batch}{vector containing potential batch effects between samples.}

\item{nthreads}{vector numeric value to indicate the threads for parallel computing .}
}
\description{
This function allows you to express your love of cats.
}
\examples{
load("simulatedData.RData")
countData = simData$abundData
dryseq_function()
}
