% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryseq_single.R
\name{dryseq_single}
\alias{dryseq_single}
\title{This function performs rhythmicity analysis for one condition}
\usage{
dryseq_single(
  countData,
  sample_name = names(countData),
  group,
  single,
  time,
  period = 24
)
}
\description{
This function performs rhythmicity analysis for one condition
}
