% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_cond.R
\name{plot_single_cond}
\alias{plot_single_cond}
\title{This function plots data and fit for one condition}
\usage{
plot_single_cond(out, gene_name)
}
\description{
This function plots data and fit for one condition
}
