% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryseq_function.R
\name{dryseq}
\alias{dryseq}
\title{Differential rhythmicity analysis for RNA-Seq datasets}
\usage{
dryseq(
  countData,
  group,
  time,
  period = 24,
  sample_name = colnames(countData),
  batch = rep("A", length(sample_name)),
  n.cores = round(detectCores() * 0.6, 0)
)
}
\arguments{
\item{countData}{matrix containing non-negative integers; each column represents a sample, each row represents a gene/transcript.}

\item{group}{vector containing the name of each sample.}

\item{time}{vector containing numeric values of the time for each sample.}

\item{period}{numeric value to indicate period length of the oscillation. Default for circadian data 24 h.}

\item{sample_name}{vector containing sample names.}

\item{batch}{vector containing potential batch effects between samples.}

\item{nthreads}{vector numeric value to indicate the threads for parallel computing .}
}
\value{
a list that contains the following data.frames: results (summary of results), parameters (rhythmic parameters), ncounts (normalized counts), counts (raw counts), cook (cook's distance)
}
\description{
This function accepts raw count data from a temporal RNA-Seq dataset of 2 or more groups. Parameters mean, phase and amplitude are given for each group.
}
\details{
test test test
}
\examples{
countData = simData$abundData
group = simData$sampleMetadata$cond
time  = simData$sampleMetadata$time
dryList = dryseq(countData,group,time)
head(dryList[["results"]])    # data frame summarizing results
head(dryList[["parameters"]]) # coefficients: phase, amplitude and mean for each group
head(dryList[["ncounts"]])    # normalized counts
head(dryList[["counts"]])     # raw counts
head(dryList[["cook"]])       # cook's distance
}
