% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryseq_function.R
\name{dryseq}
\alias{dryseq}
\title{A Cat Function}
\usage{
dryseq(
  countData,
  group,
  time,
  T_ = 24,
  sample_name = colnames(countData),
  batch = rep("A", length(sample_name)),
  n.cores = round(detectCores() * 0.6, 0)
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
