% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_models_rhythm.R
\name{plot_models_rhythm}
\alias{plot_models_rhythm}
\title{A Cat Function}
\usage{
plot_models_rhythm(DF, file_path_name, time, group, period = 24)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
