% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_models_rhythm.R
\name{plot_models_rhythm}
\alias{plot_models_rhythm}
\title{Visualization of dryseq results}
\usage{
plot_models_rhythm(DF, file_path_name, time, group, period = 24)
}
\arguments{
\item{DF}{results from dryseq}

\item{file_path_name}{folder to store output}
}
\description{
This function allows to plot the results of dryseq.
}
\examples{
XYZ
}
